practice/requirements.txt
Django==1.10.1
djangorestframework==3.4.7
httpie==0.9.6
pkg-resources==0.0.0
Pygments==2.1.3
requests==2.11.1
